First challenge completed using IPlayer code.
I'm successfully adding a delegate to the idle event of the Elevator
using TestPlayer. The entire object is cached and retrieved again. Then
the idle event is fired and the delegate is called, resulting in the
DestinationQueue filling up by three calls to GoToFloor. This is then
translated back to an ElevatorCommands object which is used to execute
on the client side.

@NogginBox  It works!!!