Fleshing out FloorController.
Look at Elevator.GoToFloor comments for where to pick up. Need to create
ElevatorCommands on Floor events, but just thinking about how to let
players access all elevators from a floor event and also how to create
ElevatorCommands from a floor event, when we don't really know which
elevator to use. We need an elevatorIndex in the ElevatorCommands.